+++ xrpl/ledger-entry-types/oracle_test.go
@@ -15,12 +15,12 @@ func TestPriceData_Flatten(t *testing.T) {
 	testcases := []struct {
 		name      string
 		priceData *PriceData
-		expected  FlatPriceData
+		expected  map[string]any
 	}{
 		{
 			name:      "pass - empty",
 			priceData: &PriceData{},
-			expected: FlatPriceData{
+			expected: map[string]any{
 				"Scale": uint8(0),
 			},
 		},
@@ -32,11 +32,84 @@ func TestPriceData_Flatten(t *testing.T) {
 				AssetPrice: 740,
 				Scale:      3,
 			},
-			expected: FlatPriceData{
+			expected: map[string]any{
 				"BaseAsset":  "XRP",
 				"QuoteAsset": "USD",
-				"AssetPrice": uint64(740),
+				"AssetPrice": "740",
 				"Scale":      uint8(3),
+			},
+		},
+		{
+			name: "pass - complete with currency more than 3 characters",
+			priceData: &PriceData{
+				BaseAsset:  "XRP",
+				QuoteAsset: "ACGBD",
+				AssetPrice: 740,
+				Scale:      3,
+			},
+			expected: map[string]any{
+				"BaseAsset":  "XRP",
+				"QuoteAsset": "ACGBD",
+				"AssetPrice": "740",
+				"Scale":      uint8(3),
+			},
+		},
+	}
+
+	for _, testcase := range testcases {
+		t.Run(testcase.name, func(t *testing.T) {
+			assert.Equal(t, testcase.priceData.Flatten(), testcase.expected)
+		})
+	}
+}
+
+func TestPriceDataWrapper_Flatten(t *testing.T) {
+	testcases := []struct {
+		name      string
+		priceData *PriceDataWrapper
+		expected  map[string]any
+	}{
+		{
+			name:      "pass - empty",
+			priceData: &PriceDataWrapper{},
+			expected:  nil,
+		},
+		{
+			name: "pass - complete",
+			priceData: &PriceDataWrapper{
+				PriceData: PriceData{
+					BaseAsset:  "XRP",
+					QuoteAsset: "USD",
+					AssetPrice: 740,
+					Scale:      3,
+				},
+			},
+			expected: map[string]any{
+				"PriceData": map[string]any{
+					"BaseAsset":  "XRP",
+					"QuoteAsset": "USD",
+					"AssetPrice": "740",
+					"Scale":      uint8(3),
+				},
+			},
+		},
+		{
+			name: "pass - complete with currency more than 3 characters",
+			priceData: &PriceDataWrapper{
+				PriceData: PriceData{
+					BaseAsset:  "XRP",
+					QuoteAsset: "ACGBD",
+					AssetPrice: 740,
+					Scale:      3,
+				},
+			},
+			expected: map[string]any{
+				"PriceData": map[string]any{
+					"BaseAsset":  "XRP",
+					"QuoteAsset": "ACGBD",
+					"AssetPrice": "740",
+					"Scale":      uint8(3),
+				},
 			},
 		},
 	}