+++ xrpl/transaction/types/domain_id.go
@@ -1,3 +1,4 @@
+//revive:disable:var-naming
 package types
 
 // DomainID returns the domain ID as a string.